fix all copy problem by variable ctrl

+++ CBoy/lib/bus.c
@@ -95,14 +95,14 @@ void bus_write(u16 address, u8 value) {
 }
 
 
-u8 bus_read16(u16 address) {
+u16 bus_read16(u16 address) {
    u16 low = bus_read(address);
    u16 high = bus_read(address + 1);
 
    return low | (high << 8);
 }
 
-void bus_write16(u16 address, u8 value) {
+void bus_write16(u16 address, u16 value) {
      bus_write(address + 1, (value >> 8) & 0xFF);
      bus_write(address, value & 0xFF);
 }
